Add list of exceptions that will trigger the retry.
This is useful when we only want to retry in very specific failures.

If a list of exceptions is provided and 'retry' > 1, we only retry if  
the exception that was raised by the example is in that list. Otherwise 
we ignore the 'retry' value and fail immediately.

If no list of exceptions is provided and 'retry' > 1, we always retry.